containers/django/requirements.txt
Django==2.2
psycopg2==2.8.6
gunicorn==20.0.4
djangorestframework==3.12.2
django-markdownx==3.0.1
pillow==8.1.1